Fix for overflowing long causing invalid json
This was caused by checking for "__json__" using PyObject_HasAttrString
which clears the error set by a previous long overflow. Thus this was dependent
on the order of processing of dict items, which explains why it was
seemingly random as the dict items are likely ordered by a hash of
the key.

This fixes GH224 and GH240.

Patch info:
    Author: Joakim Hamren <joakim.hamren@gmail.com>
    Date:   Sat Feb 4 04:21:05 2017 +0100