Following std json handling of None dict key
Previously a None dict item key would be outputted in JSON as "None".
To better align with the standard json module this was changed to output
"null". There's no proper representation of null object keys in JSON so
this is implementation specific but it seems more natural to follow
suit when it can be done without a significant performance hit.

Added and used branch prediction macros (LIKELY/UNLIKELY) as well.

Patch info:
    Author: Joakim Hamren <joakim.hamren@gmail.com>
    Date:   Sat Feb 4 16:36:14 2017 +0100